Impl. for Product

+++ src/main/java/ing/assessment/service/ProductService.java
@@ -1,10 +1,9 @@
 package ing.assessment.service;
 
-import ing.assessment.model.Product;
-
+import ing.assessment.db.Product;
 import java.util.List;
 
 public interface ProductService {
     List<Product> getAllProducts();
-    Product getProductById(Long id);
+    Product getProductById(Integer id);
 }
